- remove console log - consolidate styles - auth ui fixes - localization update
