+++ src/featureModals/SignVerify/components/Sign.tsx
@@ -25,15 +25,15 @@ import {
     MessageContainer,
     InfoIcon,
     WarningIcon,
-    Footer
-} from './style';
+    Footer,
+} from '../../style';
 
 const Sign = () => {
     const { t } = useTranslation();
     const dispatch = useDispatch();
     const { isLoading, selectedParentHash, isLedger } = useSelector((rootState: RootState) => rootState.app, shallowEqual);
     const { identities } = useSelector((rootState: RootState) => rootState.wallet, shallowEqual);
-    const { values, activeTab } = useSelector<RootState, ModalState>(state => state.modal, shallowEqual);
+    const { values, activeTab } = useSelector<RootState, ModalState>((state) => state.modal, shallowEqual);
     const { settings } = useSelector((rootState: RootState) => rootState, shallowEqual);
     const [message, setMessage] = useState('');
     const [result, setResult] = useState('');
@@ -83,7 +83,7 @@ const Sign = () => {
                 {t('components.signVerifyModal.sign_guidance')}
             </MessageContainer>
             <MainContainer>
-                <CustomTextArea label={t('general.nouns.message')} onChange={val => setMessage(val)} defaultValue={message} />
+                <CustomTextArea label={t('general.nouns.message')} onChange={(val) => setMessage(val)} defaultValue={message} />
             </MainContainer>
             <ResultContainer>
                 {error && <WarningIcon />}
